# docs/api.yaml
openapi: 3.0.3
info:
  title: 学习计划管理平台 API
  description: Personal Learning Plan Management Platform API
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /plans:
    get:
      tags: [Plans]
      summary: List learning plans
      security:
        - bearerAuth: []
      parameters:
        - name: courseId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Plan list
    post:
      tags: [Plans]
      summary: Create plan with items
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [courseId, title, startDate, endDate, items]
              properties:
                courseId:
                  type: string
                title:
                  type: string
                startDate:
                  type: string
                  format: date
                endDate:
                  type: string
                  format: date
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      title:
                        type: string
                      scheduledDate:
                        type: string
                        format: date
      responses:
        '201':
          description: Plan created
  /plans/{id}/delay:
    post:
      tags: [Plans]
      summary: Handle plan delay
      description: Marks overdue items as skipped, shifts future items forward
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Plan delayed successfully
        '400':
          description: No overdue items
...
  /wrong-answers:
    get:
      tags: [Wrong Answers]
      summary: List wrong answers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Wrong answer list
    post:
      tags: [Wrong Answers]
      summary: Add wrong answer
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Created
  /notes:
    get:
      tags: [Notes]
      summary: List notes
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Note list
    post:
      tags: [Notes]
      summary: Create note
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Created
...
  /notifications:
    get:
      tags: [Notifications]
      summary: List notifications
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Notification list
  /notifications/{id}/read:
    put:
      tags: [Notifications]
      summary: Mark notification as read
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Marked
  /notifications/read-all:
    put:
      tags: [Notifications]
      summary: Mark all as read
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All marked
...